review reformulation servlets methods and some details
