+++ src/main/java/br/com/alura/humanDev/servlet/GetAllCategoriesServlet.java
@@ -4,7 +4,6 @@ import br.com.alura.humanDev.dao.CategoryDAO;
 import br.com.alura.humanDev.entities.Category;
 import br.com.alura.humanDev.util.JPAUtil;
 
-import javax.persistence.EntityManager;
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -17,16 +16,13 @@ import java.util.List;
 @WebServlet("/listaCategorias")
 public class GetAllCategoriesServlet extends HttpServlet {
 
+    private final CategoryDAO dao = new CategoryDAO(JPAUtil.getEntityManager());
+
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws IOException, ServletException {
-
-        EntityManager em = JPAUtil.getEntityManager();
-        CategoryDAO dao = new CategoryDAO(em);
-
         List<Category> categoriesList = dao.findAllCategories();
         request.setAttribute("categories", categoriesList);
-
         RequestDispatcher rd = request.getRequestDispatcher("/categoriesList.jsp");
         rd.forward(request, response);
     }
