--- src/main/java/br/com/alura/humanDev/servlet/GetCategoryServlet.java
@@ -1,35 +1,0 @@
-package br.com.alura.humanDev.servlet;
-
-import br.com.alura.humanDev.dao.CategoryDAO;
-import br.com.alura.humanDev.entities.Category;
-import br.com.alura.humanDev.util.JPAUtil;
-
-import javax.persistence.EntityManager;
-import javax.servlet.RequestDispatcher;
-import javax.servlet.ServletException;
-import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
-import java.util.List;
-
-@WebServlet("/getCategory")
-public class GetCategoryServlet extends HttpServlet {
-
-    @Override
-    protected void doGet(HttpServletRequest request, HttpServletResponse response)
-            throws IOException, ServletException {
-
-        EntityManager em = JPAUtil.getEntityManager();
-        CategoryDAO dao = new CategoryDAO(em);
-
-        Long id = Long.valueOf(request.getParameter("id"));
-
-        Category category = dao.findCategoryById(id);
-        request.setAttribute("category", category);
-
-        RequestDispatcher rd = request.getRequestDispatcher("/putCategories.jsp");
-        rd.forward(request, response);
-    }
-}
