+++ src/main/java/br/com/alura/humanDev/servlet/PutCategoriesServlet.java
@@ -4,7 +4,6 @@ import br.com.alura.humanDev.dao.CategoryDAO;
 import br.com.alura.humanDev.entities.Category;
 import br.com.alura.humanDev.util.JPAUtil;
 
-import javax.persistence.EntityManager;
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -16,28 +15,31 @@ import java.io.IOException;
 @WebServlet("/alterarCategoria")
 public class PutCategoriesServlet extends HttpServlet {
 
+    private final CategoryDAO dao = new CategoryDAO(JPAUtil.getEntityManager());
+
+    @Override
+    protected void doGet(HttpServletRequest request, HttpServletResponse response)
+            throws IOException, ServletException {
+        Long id = Long.valueOf(request.getParameter("id"));
+        Category category = dao.findCategoryById(id);
+        request.setAttribute("category", category);
+        RequestDispatcher rd = request.getRequestDispatcher("/putCategories.jsp");
+        rd.forward(request, response);
+    }
+
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
-            throws IOException, ServletException {
-
-        EntityManager em = JPAUtil.getEntityManager();
-        CategoryDAO dao = new CategoryDAO(em);
-
-        Long id = Long.valueOf(request.getParameter("id"));
-
+            throws IOException{
         String name = request.getParameter("name");
         String code = request.getParameter("code");
         Integer ordination = Integer.valueOf(request.getParameter("ordination"));
         String description = request.getParameter("categoryDescription");
-        boolean active = Boolean.parseBoolean(request.getParameter("categoryDescription"));
+        boolean active = Boolean.parseBoolean(request.getParameter("active"));
         String icon = request.getParameter("icon");
         String color = request.getParameter("hexaColor");
 
-        Category category = new Category(id, name, code, ordination, description, active, icon, color);
+        Category category = new Category(name, code, ordination, description, active, icon, color);
         dao.update(category);
-        em.close();
-
         response.sendRedirect("listaCategorias");
-
     }
 }
